--- DZ_PARSING/src/parsing/ParserStrategy.java
@@ -1,5 +1,0 @@
-package parsing;
-
-public interface ParserStrategy {
-    public Root parse();
-}
