+++ DZ_differentParses/currency.xml
