+++ DZ_differentParses/parsers/DateParse.java
@@ -1,12 +1,11 @@
-package parsing;
+package parsers;
 
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
 import java.util.Date;
 
 public class DateParse {
-
-   public Date strToDate(String strDate) {
+    public static Date strToDate(String strDate) {
         SimpleDateFormat template = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss Z");
         Date date = null;
         try {
