+++ DZ_differentParses/parsers/ParserStratedgy.java
@@ -1,0 +1,5 @@
+package parsers;
+
+public interface ParserStratedgy {
+    public Root parse();
+}
