+++ DZ_differentParses/parsers/Root.java
@@ -1,4 +1,4 @@
-package parsing;
+package parsers;
 
 import java.util.Date;
 import java.util.List;
@@ -6,57 +6,28 @@ import java.util.List;
 public class Root {
     private String bankName;
     private String bankLocation;
-    private String dateStr;
     private Date date;
     private int baseCurrencyId;
-    private List<Currency> currency;
-
-    public String getBankName() {
-        return bankName;
-    }
+    private List<Currency> currencies;
 
     public void setBankName(String bankName) {
         this.bankName = bankName;
-    }
-
-    public String getBankLocation() {
-        return bankLocation;
     }
 
     public void setBankLocation(String bankLocation) {
         this.bankLocation = bankLocation;
     }
 
-
-    public Date getDate() {
-        return date;
-    }
-
     public void setDate(Date date) {
         this.date = date;
-    }
-
-    public int getBaseCurrencyId() {
-        return baseCurrencyId;
     }
 
     public void setBaseCurrencyId(int baseCurrencyId) {
         this.baseCurrencyId = baseCurrencyId;
     }
 
-    public List<Currency> getCurrencies() {
-        return currency;
-    }
-
     public void setCurrencies(List<Currency> currencies) {
-        this.currency = currencies;
-    }
-    public void setDateStr(String dateStr) {
-        this.dateStr = dateStr;
-    }
-
-    public String getDateStr() {
-        return dateStr;
+        this.currencies = currencies;
     }
 
     @Override
@@ -64,9 +35,9 @@ public class Root {
         return "Root{" +
                 "bankName='" + bankName + '\'' +
                 ", bankLocation='" + bankLocation + '\'' +
-                ", date=" + dateStr +
+                ", date=" + date +
                 ", baseCurrencyId=" + baseCurrencyId +
-                ", currency=" + currency +
+                ", currency=" + currencies +
                 '}';
     }
 }
